[fix] #42 modify annotation conditional statement for permission confirmation (#44)

+++ src/main/java/org/kkumulkkum/server/annotation/IsMember.java
@@ -8,6 +8,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface IsMember {
-    long meetingIdParamIndex() default -1;
-    long promiseIdParamIndex() default -1;
+    int meetingIdParamIndex() default -1;
+    int promiseIdParamIndex() default -1;
 }
